+++ src/main/java/com/example/yanolja/domain/review/controller/ReviewControllerAdvice.java
@@ -1,7 +1,6 @@
 package com.example.yanolja.domain.review.controller;
 
-import com.example.yanolja.domain.review.error.ReviewNotFoundException;
-import com.example.yanolja.domain.review.error.ReviewOperationException;
+import com.example.yanolja.domain.review.exception.ReviewNotFoundException;
 import com.example.yanolja.global.util.ResponseDTO;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
@@ -19,14 +18,5 @@ public class ReviewControllerAdvice {
                 exception.getMessage()));
     }
 
-    @ExceptionHandler(value = {
-        ReviewOperationException.class
-    })
-    public ResponseEntity<ResponseDTO<Object>> handleReviewOperationException(
-        ReviewOperationException exception
-    ) {
-        return ResponseEntity.status(exception.getErrorCode().getHttpStatus()).body(
-            ResponseDTO.res(exception.getErrorCode().getHttpStatus(),
-                exception.getMessage()));
-    }
+
 }
