+++ src/main/java/com/example/yanolja/domain/review/exception/ReviewNotFoundException.java
@@ -1,11 +1,14 @@
-package com.example.yanolja.domain.review.error;
+package com.example.yanolja.domain.review.exception;
 
 import com.example.yanolja.global.exception.ApplicationException;
 import com.example.yanolja.global.exception.ErrorCode;
 
 public class ReviewNotFoundException extends ApplicationException {
 
-    public ReviewNotFoundException(ErrorCode errorCode) {
-        super(errorCode);
+    private static final ErrorCode ERROR_CODE = ErrorCode.REVIEW_NOT_FOUND;
+
+    public ReviewNotFoundException() {
+
+        super(ERROR_CODE);
     }
 }
